Avoid exposing environmental variables
